remove clj-browserchannel-netty-adapter
it was horribly out of date, didn't even build (broken dependencies),
and when i briefly looked into fixing it up, i quickly realized that
aleph/netty has changed quite a bit in the few years since this adapter
was last touched. i really don't know anything about aleph/netty as
i've never used it so i don't think it's a good idea for me to be trying
to update this.

ultimately i think removing this is the best option for now rather then
keeping it around in a severly out of date and broken state.